refactor: flat and reorganize sub-modules
