+++ examples/agent_execution/agent.py
@@ -1,10 +1,11 @@
 from peak import Agent
 from peak.behaviours import OneShotBehaviour
 
+
 class agent(Agent):
     class HelloWorld(OneShotBehaviour):
         async def run(self) -> None:
-            print('Hello World')
+            print("Hello World")
             await self.agent.stop()
 
     async def setup(self) -> None:
