+++ peak/behaviours.py
@@ -1,107 +1,10 @@
 import logging
-from abc import ABCMeta, abstractmethod
 from json import dumps as json_dumps
 from typing import Callable
 
-import aioxmpp
-import spade.behaviour
-from aioxmpp import JID
-from spade.message import Message
-
-import peak
-from peak import Agent
-
-aioxmpp.pubsub.xso.as_payload_class(aioxmpp.Message)
-
-
-def slice_jid(jid):
-    jid = JID.fromstr(jid)
-    pubsub = JID.fromstr(jid.domain)
-    node = jid.localpart
-    return pubsub, node
-
-def df_jid(domain):
-    return 'df@'+domain+'/admin'
-
-class _Behaviour:
-
-    agent: Agent
-
-    async def send_to_group(self, msg: Message):
-        """Sends a message to a group chat.
-
-        When sending a message to a group the agent joins the group first. The parameter
-        'leave' tells the method if the agent leaves, or not, the group after sending the
-        message. (If the intention is to send a single request to a new group the best
-        option would be to leave the group chat, if the intention is to send a message
-        to a group wich the agent already belongs to, it's better to not leave)
-        Args:
-            msg (mas.Message): The Message.
-            group (str, optional): Name of the group to send the message to. If None is given the
-                                   the message is sent to the MAS group. Defaults to None.
-            leave (bool, optional): If true, agent leaves the group after sending the message. Defaults to False.
-        """
-        raw_msg = msg.prepare()
-        try:
-            await self.agent.groups[str(msg.to)].send_message(raw_msg)
-        except:
-            room, future = self.agent.muc_client.join(msg.to, self.agent.name)
-            await future
-            await room.send_message(raw_msg)
-            await room.leave()
-
-    async def change_node_affiliations(self, jid: str, affiliations_to_set: tuple):
-        """Changes PubSub node affiliations.
-
-        Args:
-            jid (str): JID of the node, e.g. node@pubsub.example.com
-            affiliations_to_set (tuple[str,str]): each tuple must contain the JID of the user and the affiliation (e.g.'owner','publisher')
-        """
-        pubsub, node = slice_jid(jid)
-        user, aff = affiliations_to_set
-        await self.agent.pubsub_client.change_node_affiliations(
-            pubsub, node, [(JID.fromstr(user), aff)]
-        )
-
-    async def subscribe(self, jid: str):
-        pubsub, node = slice_jid(jid)
-        await self.agent.pubsub_client.subscribe(pubsub, node)
-        await self.agent.pubsub_client.on_item_published.connect(self.on_item_published)
-
-    async def publish(self, msg: Message):
-        jid = msg.to.domain
-        node = msg.to.localpart
-        await self.agent.pubsub_client.publish(JID.fromstr(jid), node, msg.prepare())
-
-    async def notify(self, jid: str):
-        pubsub, node = slice_jid(jid)
-        await self.agent.pubsub_client.notify(pubsub, node)
-
-    async def unsubscribe(self, jid: str):
-        pubsub, node = slice_jid(jid)
-        await self.agent.pubsub_client.unsubscribe(pubsub, node)
-
-    @abstractmethod
-    def on_item_published(jid, node, item, *, message=None):
-        pass
-
-
-class OneShotBehaviour(spade.behaviour.OneShotBehaviour, _Behaviour, metaclass=ABCMeta):
-    pass
-
-
-class PeriodicBehaviour(
-    spade.behaviour.PeriodicBehaviour, _Behaviour, metaclass=ABCMeta
-):
-    pass
-
-
-class CyclicBehaviour(spade.behaviour.CyclicBehaviour, _Behaviour, metaclass=ABCMeta):
-    pass
-
-
-class FSMBehaviour(spade.behaviour.FSMBehaviour, _Behaviour, metaclass=ABCMeta):
-    pass
+from peak import Message, SyncAgent, Template, DF
+from peak.core import CyclicBehaviour, OneShotBehaviour, PeriodicBehaviour
+from peak.properties import Property
 
 
 class JoinGroup(OneShotBehaviour):
@@ -112,8 +15,8 @@ class JoinGroup(OneShotBehaviour):
         self.tags = tags
 
     async def run(self):
-        msg = peak.Message()
-        msg.to = df_jid(self.agent.jid.domain)
+        msg = Message()
+        msg.to = DF.name(self.agent.jid.domain)
         msg.set_metadata("resource", "treehierarchy")
         msg.set_metadata("path", self.path)
         msg.set_metadata("domain", self.domain)
@@ -133,8 +36,8 @@ class LeaveGroup(OneShotBehaviour):
         self.domain = domain
 
     async def run(self):
-        msg = peak.Message()
-        msg.to = df_jid(self.agent.jid.domain)
+        msg = Message()
+        msg.to = DF.name(self.agent.jid.domain)
         msg.set_metadata("resource", "treehierarchy")
         msg.set_metadata("path", self.path)
         msg.set_metadata("domain", self.domain)
@@ -157,8 +60,8 @@ class SearchGroup(OneShotBehaviour):
         self.kargs = kargs
 
     async def run(self):
-        msg = peak.Message()
-        msg.to = df_jid(self.agent.jid.domain)
+        msg = Message()
+        msg.to = DF.name(self.agent.jid.domain)
         msg.set_metadata("resource", "searchgroup")
         msg.set_metadata("tags", str(self.tags))
         await self.send(msg)
@@ -194,7 +97,7 @@ class ExportData(OneShotBehaviour):
 
     async def run(self) -> None:
         logger = logging.getLogger(self.__class__.__name__)
-        if isinstance(self.agent, peak.SyncAgent):
+        if isinstance(self.agent, SyncAgent):
             logger.debug("synchronizer detected")
             self.agent.add_behaviour(
                 _ExportDataSync(
@@ -238,9 +141,9 @@ class _ExportDataSync(CyclicBehaviour):
     async def on_start(self) -> None:
         self.logger = logging.getLogger(self.__class__.__name__)
         self.logger.info("behaviour initiated")
-        template_step = peak.Template()
+        template_step = Template()
         template_step.set_metadata("sync", "step")
-        template_stop = peak.Template()
+        template_stop = Template()
         template_stop.set_metadata("sync", "stop")
         template = template_step | template_stop
         self.set_template(template)
@@ -249,7 +152,7 @@ class _ExportDataSync(CyclicBehaviour):
         for property in self.properties:
             self.file_data[property] = []
         if self.to_graph:
-            msg = peak.Message(to=df_jid(self.agent.jid.domain))
+            msg = Message(to=DF.name(self.agent.jid.domain))
             msg.body = "Create graph " + self.graph_name
             msg.metadata = {
                 "resource": "graph",
@@ -268,14 +171,14 @@ class _ExportDataSync(CyclicBehaviour):
                 current_data = dict()
                 for property in self.file_data:
                     attribute = getattr(self.agent, property)
-                    if type(attribute) is peak.Property:
+                    if type(attribute) is Property:
                         self.file_data[property].append(attribute.current_value)
                         current_data[property] = attribute.current_value
                     else:
                         self.file_data[property].append(attribute)
                         current_data[property] = attribute
                 if self.to_graph:
-                    msg = peak.Message(to=df_jid(self.agent.jid.domain))
+                    msg = Message(to=DF.name(self.agent.jid.domain))
                     msg.body = "Update graph " + self.file_name
                     msg.metadata = {
                         "resource": "graph",
@@ -314,7 +217,7 @@ class _ExportData(PeriodicBehaviour):
         for property in self.properties:
             self.file_data[property] = []
         if self.to_graph:
-            msg = peak.Message(to=df_jid(self.agent.jid.domain))
+            msg = Message(to=DF.name(self.agent.jid.domain))
             msg.body = "Create graph " + self.graph_name
             msg.metadata = {
                 "resource": "graph",
@@ -330,7 +233,7 @@ class _ExportData(PeriodicBehaviour):
         current_data = dict()
         for property in self.file_data:
             attribute = getattr(self.agent, property)
-            if type(attribute) is peak.Property:
+            if type(attribute) is Property:
                 self.file_data[property].append(attribute.current_value)
                 current_data[property] = attribute.current_value
             else:
@@ -340,7 +243,7 @@ class _ExportData(PeriodicBehaviour):
                 'exporting property "' + property + '": ' + str(current_data[property])
             )
         if self.to_graph:
-            msg = peak.Message(to=df_jid(self.agent.jid.domain))
+            msg = Message(to=DF.name(self.agent.jid.domain))
             msg.body = "Update graph " + self.graph_name
             msg.metadata = {
                 "resource": "graph",
